Refactor: home 화면 파일 구조 개선 및 import 경로 수정
