+++ src/app/home/components/index.ts
@@ -1,0 +1,7 @@
+// 모든 컴포넌트들을 깔끔하게 re-export
+export * from './traits';
+export * from './layout';
+export * from './seo';
+
+// 메인 컴포넌트
+export { ClientAnimations } from './ClientAnimations';
