+++ src/app/home/components/layout/HeroSection.tsx
