+++ src/app/home/components/layout/ScrollPageIndicator.tsx
