+++ src/app/home/components/layout/ThankYouSection.tsx
