+++ src/app/home/components/layout/index.ts
@@ -1,0 +1,5 @@
+// Layout 관련 컴포넌트들
+export { HeroSection } from './HeroSection';
+export { ThankYouSection } from './ThankYouSection';
+export { LoadingScreen } from './LoadingScreen';
+export { ScrollProgressIndicator } from './ScrollPageIndicator';
