+++ src/app/home/components/seo/I18nStaticContent.tsx
