+++ src/app/home/components/seo/index.ts
@@ -1,0 +1,3 @@
+// SEO 관련 컴포넌트들
+export { I18nStaticContent } from './I18nStaticContent';
+export { StructuredData } from './StructuredData';
