+++ src/app/home/components/traits/ContactSection.tsx
