+++ src/app/home/components/traits/TabContent.tsx
