+++ src/app/home/components/traits/TraitCard.tsx
