+++ src/app/home/components/traits/TraitCardList.tsx
