+++ src/app/home/components/traits/index.ts
@@ -1,0 +1,7 @@
+// Traits 관련 컴포넌트들
+export { TraitSection } from './TraitSection';
+export { default as TraitCard } from './TraitCard';
+export { TraitCardList } from './TraitCardList';
+export { TabButton } from './TabButton';
+export { TabContent } from './TabContent';
+export { ContactSection } from './ContactSection';
